requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
httpx>=0.28
pydantic-settings>=2.7
python-multipart>=0.0.18
python-dateutil>=2.9
gpxpy>=1.6
itsdangerous>=2.2
starlette>=0.45
